requirements.txt
adal==1.2.7
azure-batch==12.0.0
azure-common==1.1.28
azure-core==1.26.1
azure-storage-blob==12.8.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
idna==3.4
isodate==0.6.1
msrest==0.7.1
msrestazure==0.6.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
